+++ NEWS.md
@@ -1,3 +1,9 @@
+# mlr3extralearners 0.5.3
+
+* Fixed bugs in catboost for classification
+* Removed factor feature types from catboost
+* Added `install_catboost` to make installation from CatBoost simpler
+
 # mlr3extralearners 0.5.2
 
 * Fixed learner tests